requirements.txt
streamlit
google-auth
google-auth-oauthlib
pandas
numpy
yfinance
plotly
matplotlib
